:recycle: refactor: Using common struct response for admin & app & format style code
refactor: Using common struct response for admin & app & format style code
