+++ docs/app/docs.go
@@ -139,7 +139,7 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/responsemodel.ResponseCreate"
+                            "$ref": "#/definitions/response.ResponseCreate"
                         }
                     }
                 }
@@ -217,7 +217,7 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/responsemodel.ResponseUpdate"
+                            "$ref": "#/definitions/response.ResponseUpdate"
                         }
                     }
                 }
@@ -301,7 +301,7 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/responsemodel.ResponseCreate"
+                            "$ref": "#/definitions/response.ResponseCreate"
                         }
                     }
                 }
@@ -379,7 +379,7 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/responsemodel.ResponseUpdate"
+                            "$ref": "#/definitions/response.ResponseUpdate"
                         }
                     }
                 }
@@ -574,7 +574,7 @@ const docTemplate = `{
                 }
             }
         },
-        "responsemodel.ResponseCreate": {
+        "response.ResponseCreate": {
             "type": "object",
             "properties": {
                 "_id": {
@@ -582,7 +582,7 @@ const docTemplate = `{
                 }
             }
         },
-        "responsemodel.ResponseUpdate": {
+        "response.ResponseUpdate": {
             "type": "object",
             "properties": {
                 "_id": {
